feat(invoices): make the PDF modal the single export entry point
The list row had both a "view PDF" button and a separate "download PDF"
format dropdown next to it. Consolidate: the row's PDF button (now
"Export PDF") opens a modal that both previews the PDF fit to the whole
page and offers the full format list (PDF Standard/Factur-X/ZUGFeRD/
XRechnung/UBL/CII + XML and plugin formats) for download.

- invoice-pdf-view.tsx: preview opens at page-fit zoom; header has a
  right-aligned format dropdown reusing the list's download machinery.
- invoice-list.tsx: drop the standalone download dropdown and its
  trigger/state/effect/imports; rename the PDF button tooltip viewPdf
  -> exportPdf.
- i18n: rename invoices.list.tooltips.viewPdf to exportPdf in all
  locales (downloadPdf/actions keys kept, now used by the modal).
