+++ frontend/src/pages/(app)/invoices/_components/invoice-list.tsx
@@ -1,8 +1,7 @@
-import { Code, Download, Edit, FileText, Mail, Plus, ReceiptText as PaymentText, Search, Trash2 } from "lucide-react"
+import { Edit, Mail, Plus, ReceiptText as PaymentText, Search, Trash2 } from "lucide-react"
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
-import { forwardRef, useEffect, useImperativeHandle, useState } from "react"
-import { useGet, useGetRaw, usePost } from "@/hooks/use-fetch"
+import { forwardRef, useImperativeHandle, useState } from "react"
+import { usePost } from "@/hooks/use-fetch"
 
 import BetterPagination from "../../../../components/pagination"
 import { Badge } from "@/components/ui/badge"
@@ -41,18 +40,12 @@ export interface InvoiceListHandle {
     handleAddClick: () => void
 }
 
-interface PluginPdfFormat {
-    format_name: string
-    format_key: string
-}
-
 export const InvoiceList = forwardRef<InvoiceListHandle, InvoiceListProps>(
     (
         { invoices, loading, title, description, searchTerm, onSearchChange, statusFilter, onStatusFilterChange, statusCounts, page, pageCount, setPage, mutate, emptyState, showCreateButton = false, onAddClick },
         ref,
     ) => {
         const { t } = useTranslation()
-        const { data: pdf_formats } = useGet<PluginPdfFormat[]>('/api/plugins/formats')
         const { trigger: triggerSendInvoiceByEmail, loading: sendInvoiceByEmailLoading } = usePost(`/api/invoices/send`)
 
         const [createInvoiceDialog, setCreateInvoiceDialog] = useState<boolean>(false)
@@ -61,42 +54,12 @@ export const InvoiceList = forwardRef<InvoiceListHandle, InvoiceListProps>(
         const [viewInvoicePdfDialog, setViewInvoicePdfDialog] = useState<Invoice | null>(null)
         const [deleteInvoiceDialog, setDeleteInvoiceDialog] = useState<Invoice | null>(null)
         const [sendInvoiceDialog, setSendInvoiceDialog] = useState<Invoice | null>(null)
-        const [downloadTrigger, setDownloadTrigger] = useState<{
-            invoice: Invoice
-            format: string
-            file_format: 'pdf' | 'xml'
-            id: number
-        } | null>(null)
-
-        const { data: file } = useGetRaw<Response>(
-            downloadTrigger
-                ? `/api/invoices/${downloadTrigger.invoice.id}/download/${downloadTrigger.file_format}?format=${downloadTrigger.format}`
-                : null,
-        )
 
         useImperativeHandle(ref, () => ({
             handleAddClick() {
                 setCreateInvoiceDialog(true)
             },
         }))
-
-        useEffect(() => {
-            if (downloadTrigger && file) {
-                file.arrayBuffer().then((buffer) => {
-                    const blob = new Blob([buffer], { type: `application/${downloadTrigger.file_format}` })
-                    const url = URL.createObjectURL(blob)
-                    const link = document.createElement("a")
-                    link.href = url
-                    link.download = `invoice-${downloadTrigger.invoice.number}-${downloadTrigger.format}.${downloadTrigger.file_format}`
-                    document.body.appendChild(link)
-                    link.click()
-                    document.body.removeChild(link)
-                    URL.revokeObjectURL(url)
-                    setDownloadTrigger(null) // Reset
-                }).catch(() => {
-                })
-            }
-        }, [downloadTrigger, file])
 
         function handleEdit(invoice: Invoice) {
             setEditInvoiceDialog(invoice)
@@ -112,10 +75,6 @@ export const InvoiceList = forwardRef<InvoiceListHandle, InvoiceListProps>(
 
         function handleDelete(invoice: Invoice) {
             setDeleteInvoiceDialog(invoice)
-        }
-
-        function handleDownload({ invoice, format, file_format }: { invoice: Invoice; format: string; file_format: 'pdf' | 'xml' }) {
-            setDownloadTrigger({ invoice, format, file_format, id: Date.now() })
         }
 
         const getStatusColor = (status: string) => {
@@ -337,7 +296,7 @@ export const InvoiceList = forwardRef<InvoiceListHandle, InvoiceListProps>(
                                             {invoice.status !== InvoiceStatus.UPCOMING && (
                                             <div className="grid grid-cols-2 lg:flex justify-start sm:justify-end gap-1 md:gap-2">
                                                 <Button
-                                                    tooltip={t("invoices.list.tooltips.viewPdf")}
+                                                    tooltip={t("invoices.list.tooltips.exportPdf")}
                                                     variant="ghost"
                                                     size="icon"
                                                     onClick={() => handleViewPdf(invoice)}
@@ -345,65 +304,6 @@ export const InvoiceList = forwardRef<InvoiceListHandle, InvoiceListProps>(
                                                 >
                                                     <PaymentText className="h-4 w-4" />
                                                 </Button>
-
-                                                <DropdownMenu>
-                                                    <DropdownMenuTrigger asChild>
-                                                        <Button
-                                                            tooltip={t("invoices.list.tooltips.downloadPdf")}
-                                                            variant="ghost"
-                                                            size="icon"
-                                                            className="text-gray-600 hover:text-amber-600"
-                                                        >
-                                                            <Download className="h-4 w-4" />
-                                                        </Button>
-                                                    </DropdownMenuTrigger>
-                                                    <DropdownMenuContent align="center" className="[&>*]:cursor-pointer w-48">
-                                                        <DropdownMenuLabel className="flex items-center gap-2 text-xs font-medium text-muted-foreground">
-                                                            <FileText className="h-3 w-3" />
-                                                            {t("invoices.list.actions.downloadPdf")}
-                                                        </DropdownMenuLabel>
-
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "", file_format: "pdf" })}>Standard</DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "facturx", file_format: "pdf" })}>Factur-X</DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "zugferd", file_format: "pdf" })}>ZUGFeRD</DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "xrechnung", file_format: "pdf" })}>
-                                                            XRechnung
-                                                        </DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "ubl", file_format: "pdf" })}>UBL</DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "cii", file_format: "pdf" })}>CII</DropdownMenuItem>
-
-                                                        <DropdownMenuSeparator />
-
-                                                        <DropdownMenuLabel className="flex items-center gap-2 text-xs font-medium text-muted-foreground">
-                                                            <Code className="h-3 w-3" />
-                                                            {t("invoices.list.actions.downloadXml")}
-                                                        </DropdownMenuLabel>
-
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "facturx", file_format: "xml" })}>
-                                                            Factur-X
-                                                        </DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "zugferd", file_format: "xml" })}>
-                                                            ZUGFeRD
-                                                        </DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "xrechnung", file_format: "xml" })}>
-                                                            XRechnung
-                                                        </DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "ubl", file_format: "xml" })}>
-                                                            UBL
-                                                        </DropdownMenuItem>
-                                                        <DropdownMenuItem onClick={() => handleDownload({ invoice, format: "cii", file_format: "xml" })}>
-                                                            CII
-                                                        </DropdownMenuItem>
-                                                        {pdf_formats?.map((format) => (
-                                                            <DropdownMenuItem
-                                                                key={format.format_key}
-                                                                onClick={() => handleDownload({ invoice, format: format.format_key, file_format: "xml" })}
-                                                            >
-                                                                {format.format_name}
-                                                            </DropdownMenuItem>
-                                                        ))}
-                                                    </DropdownMenuContent>
-                                                </DropdownMenu>
 
                                                 {invoice.status !== "PAID" && (
                                                     <Button
